Fix Memory Leak in Capsule Debug Library
[Issue Description]
CapsuleDebugLib allocate a huge buffer for capsule log, but the
default setting of this feature is disabled and it will not be used.

[Resolution]
Change the definition of the capsule log buffer in the protocol to
a pointer.
Allocate memory for capsule log buffer when the function is enabled.

Package/Module:
CapsuleDebugLib

[Impacted Platform]
OEMx

Hsd-es-id: 13012704425
Original commit date: Fri Feb 21 14:46:08 2025 +0800
Change-Id: I567c43163d76fd5b799589dd97b55735e5834d77
Original commit hash: 6ffad210f4666f555ee35e1bd36033a1d31ecdf3