PlatSamplePkg:Remove incorrect PchXhciUaolEnable programming
[Issue Description]
PchXhciUaolEnable is the UPD for USB Audio Offload in USB
config in XHCI, always update it to TRUE with UAOL config
in TCSS is not correct.

[Resolution]
Remove the UPD PchXhciUaolEnable update logic in UpdatePeiSaPolicy
As this UPD is not for the UAOL config in TCSS.
And UAOL is not required in TCSS, so set default to FALSE in TCSS
config.

Package/Module:
PlatSamplePkg/PeiPolicyUpdateLib

[Impacted Platform]
All

Hsd-es-id: 15017478427
Original commit date: Thu Feb 27 02:20:35 2025 +0000
Change-Id: I9444f73e0727b7782f1e8635a6b817d993103368
Original commit hash: 0855cf667b1631d7fe2b91f1b5135b18b27cff8e