Improve Build Containerization by Considering CPU Utilization Quotas
[Issue Description]
When running in a container, it is possible for CPU utilization quotas to be
imposed. This prevents a single container from using all available CPU time.

One cannot blindly assume that the number of logical CPU cores that exist on
the system is equal to the number of concurrent threads one is allowed to
execute on at a time.

[Resolution]
The fix is to check if a CPU utilization quota has been imposed. If there is
a quota, then limit the concurrent thread count to fit inside of the quota.
If here is no quota, then all available CPU threads can be used.

Package/Module: OneSiliconPkg, FspPkg, BoardPkg, OpenBoardPkg

[Impacted Platform]
ALL

Hsd-es-id: 22021211849
Original commit date: Thu May 1 15:24:03 2025 -0700
Change-Id: Ic79f8c796421c48667a2d108e151b3cbfe8aaee3
Original commit hash: 4440baf5589df95ed0283fbef5439a7832ea251f